AI Polish 2/27 10:22pm

+++ TetraiderEngine/Source/AI_DashingEngage.h
@@ -28,7 +28,7 @@ public:
 	virtual void HandleEvent(Event* pEvent);
 
 private:
-	int engageTimer;
+	float engageTimer;
 	float sinceEngage;
 	
 };
